+++ plugins/woocommerce/tests/e2e/env/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added quotes around `WORDPRESS_TITLE` value in .env file to address issue with docker compose 2 "key cannot contain a space" error.
 - Added `LATEST_WP_VERSION_MINUS` that allows setting a number to subtract from the current WordPress version for the WordPress Docker image.
+- Swapped out `@Automattic/puppeteer-utils` for `@lanej0/playwright-utils`
 
 # 0.2.3